requirements.txt
Django==4.2.7 
Pillow==10.0.1 
django-crispy-forms==2.0 
crispy-bootstrap5==0.7 
django-ckeditor==6.7.0 
python-decouple==3.8 
psycopg2-binary==2.9.7 
mysqlclient==2.2.0 
django-cors-headers==4.3.1 
django-ratelimit==4.1.0 
django-storages==1.14.2 
boto3==1.29.7 
django-modeltranslation==0.18.11 
django-compressor==4.4 
django-debug-toolbar==4.2.0 
djangorestframework==3.14.0 
django-allauth==0.57.0 
folium==0.15.0 
pytz==2023.3 
gunicorn==21.2.0 
whitenoise==6.6.0 
celery==5.3.4 
redis==5.0.1 
django-extensions==3.2.3 
ipython==8.16.1